refactor: toc로 폴더 이름 변경 (#447)
