+++ src/layouts/PostLayout.tsx
@@ -16,8 +16,8 @@ import TagInPost from '@/components/blog/tag-in-post';
 import ViewCountUpdater from '@/components/blog/view-count-updater';
 import Comments from '@/components/comments';
 import ScrollTopAndBottom from '@/components/common/scroll-top-and-bottom';
-import LargeWidthToc from '@/components/TOC/large-width-toc';
-import SmallWidthToc from '@/components/TOC/small-width-toc';
+import LargeWidthToc from '@/components/toc/large-width-toc';
+import SmallWidthToc from '@/components/toc/small-width-toc';
 
 interface Props {
   children: ReactNode;
